+++ tests/urls.py
@@ -3,9 +3,11 @@
 from django.conf import settings
 from django.conf.urls.static import static
 from django.contrib import admin
-from django.urls import re_path
+from django.urls import path, re_path
 
 from rest_framework.routers import DefaultRouter
+
+from drf_spectacular.views import SpectacularSwaggerView
 
 from .fake_app.api import views
 
@@ -37,3 +39,10 @@ if settings.DEBUG:
         settings.STATIC_URL,
         document_root=settings.STATIC_ROOT,
     )
+    urlpatterns += [
+        path(
+            "api/schema/swagger-ui/",
+            SpectacularSwaggerView.as_view(url_name="schema"),
+            name="swagger-ui",
+        ),
+    ]